Don't call assetMap.Keys() on each call to assetPath holy shit.
This will allocate a slice and fill with every value in the manifest
then sort it. So if you have 1k items in the manifest this is slow as hell.

I've also changed it so we only reload the manifest in dev mode once.
Not 9 billion times per a page load.

This brings my page load time down from 16ms to 4ms... Jesus.
This slow down is also in production since Keys() always called.